+++ R/model.R
@@ -413,13 +413,26 @@ fit <- function(object, x = NULL, y = NULL, batch_size=NULL, epochs=10,
     initial_epoch = as.integer(initial_epoch)
   )
   
-  if (!is.null(validation_data))
-    args$validation_data <- keras_array(validation_data)
-  
-  if (!is.null(x))
-    args$x <- keras_array(x)
-  if (!is.null(y))
-    args$y <- keras_array(y)
+  # resolve validation_Data (check for TF dataset)
+  if (!is.null(validation_data)) {
+    dataset <- resolve_tensorflow_dataset(validation_data)
+    if (!is.null(dataset))
+      args$validation_data <- dataset
+    else
+      args$validation_data <- keras_array(validation_data)  
+  }
+    
+  # resolve x and y (check for TF dataset)
+  dataset <- resolve_tensorflow_dataset(x)
+  if (!is.null(dataset)) {
+    args$x <- dataset[[1]]
+    args$y <- dataset[[2]]
+  } else {
+    if (!is.null(x))
+      args$x <- keras_array(x)
+    if (!is.null(y))
+      args$y <- keras_array(y) 
+  }
   
   if (keras_version() >= "2.0.7") {
     args$steps_per_epoch <- as_nullable_integer(steps_per_epoch)
@@ -472,14 +485,23 @@ evaluate.keras.engine.training.Model <- function(object, x = NULL, y = NULL, bat
   
   # args
   args <- list(
-    x = keras_array(x),
-    y = keras_array(y),
     batch_size = as_nullable_integer(batch_size),
     verbose = as.integer(verbose),
     sample_weight = sample_weight
   )
+  
+  # resolve x and y (check for TF dataset)
+  dataset <- resolve_tensorflow_dataset(x)
+  if (!is.null(dataset)) {
+    args$x <- dataset[[1]]
+    args$y <- dataset[[2]]
+  } else {
+    args$x <- keras_array(x)
+    args$y <- keras_array(y) 
+  }
+  
   if (keras_version() >= "2.0.7")
-    args$steps <- steps
+    args$steps <- as_nullable_integer(steps)
   
   # perform evaluation
   result <- do.call(object$evaluate, args)
@@ -523,12 +545,20 @@ predict.keras.engine.training.Model <- function(object, x, batch_size=NULL, verb
   
   # args
   args <- list(
-    x = keras_array(x), 
     batch_size = as_nullable_integer(batch_size),
     verbose = as.integer(verbose)
   )
+  
+  # resolve x (check for TF dataset)
+  dataset <- resolve_tensorflow_dataset(x)
+  if (!is.null(dataset)) {
+    args$x <- dataset[[1]]
+  } else {
+    args$x <- keras_array(x)
+  }
+  
   if (keras_version() >= "2.0.7")
-    args$steps <- steps
+    args$steps <- as_nullable_integer(steps)
   
   # call predict
   do.call(object$predict, args)
@@ -550,11 +580,19 @@ predict.keras.engine.training.Model <- function(object, x, batch_size=NULL, verb
 #' 
 #' @export
 predict_proba <- function(object, x, batch_size = NULL, verbose = 0, steps = NULL) {
+  
   args <- list(
-    x = keras_array(x),
     batch_size = as_nullable_integer(batch_size),
     verbose = as.integer(verbose)
   )
+  
+  # resolve x (check for TF dataset)
+  dataset <- resolve_tensorflow_dataset(x)
+  if (!is.null(dataset)) {
+    args$x <- dataset[[1]]
+  } else {
+    args$x <- keras_array(x)
+  }
   
   if (keras_version() >= "2.1.3")
     args$steps <- as_nullable_integer(steps)
@@ -566,11 +604,18 @@ predict_proba <- function(object, x, batch_size = NULL, verbose = 0, steps = NUL
 #' @export
 predict_classes <- function(object, x, batch_size = NULL, verbose = 0, steps = NULL) {
   args <- list(
-    x = keras_array(x),
     batch_size = as_nullable_integer(batch_size),
     verbose = as.integer(verbose)
   )
-
+  
+  # resolve x (check for TF dataset)
+  dataset <- resolve_tensorflow_dataset(x)
+  if (!is.null(dataset)) {
+    args$x <- dataset[[1]]
+  } else {
+    args$x <- keras_array(x)
+  }
+  
   if (keras_version() >= "2.1.3")
     args$steps <- as_nullable_integer(steps)
   
@@ -896,6 +941,33 @@ is_main_thread_generator.keras.preprocessing.image.Iterator <- function(x) {
     FALSE
   }
 }
+
+is_tensorflow_dataset <- function(x) {
+  inherits(x, "tensorflow.python.data.ops.dataset_ops.Dataset")
+}
+
+resolve_tensorflow_dataset <- function(x) {
+  
+  if (is_tensorflow_dataset(x)) {
+    
+    # check version compatibility
+    if (keras_version() < "2.2.0")
+      stop("Keras v2.2 or higher is required for direct tensor input to models", call. = FALSE)
+    if (!is_backend("tensorflow"))
+      stop("The tensorflow backend is required for direct tensor input to models", call. = FALSE)
+    if (tensorflow::tf_version() < "1.8")
+      stop("TensorFlow v1.8 or higher is required for direct tensor input to models", call. = FALSE)
+    
+    # yield iterators
+    iter = x$make_one_shot_iterator()
+    iter$get_next()
+    
+  } else {
+    NULL
+  }
+}
+
+
 
   
 #' Retrieves a layer based on either its name (unique) or index.